Revert "make settings, charger and bluetooth page"
