--- app/src/main/res/drawable/blackcar.xml
@@ -1,15 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:autoMirrored="true"
-    android:height="28dp"
-    android:tint="#101010"
-    android:viewportHeight="24"
-    android:viewportWidth="24"
-    android:width="28dp">
-      
-    <path android:fillColor="@android:color/white"
-        android:pathData="M18.92,2.01C18.72,1.42 18.16,1 17.5,1h-11C5.84,1 5.29,1.42 5.08,2.01L3,8v8c0,0.55 0.45,1 1,1h1c0.55,0 1,-0.45 1,-1v-1h12v1c0,0.55 0.45,1 1,1h1c0.55,0 1,-0.45 1,-1V8L18.92,2.01zM6.5,12C5.67,12 5,11.33 5,10.5S5.67,9 6.5,9S8,9.67 8,10.5S7.33,12 6.5,12zM17.5,12c-0.83,0 -1.5,-0.67 -1.5,-1.5S16.67,9 17.5,9S19,9.67 19,10.5S18.33,12 17.5,12zM5,7l1.5,-4.5h11L19,7H5z"/>
-      
-    <path android:fillColor="@android:color/white"
-        android:pathData="M7,20l4,0l0,-2l6,3l-4,0l0,2z"/>
-    
-</vector>
