--- app/src/main/res/drawable/charge.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:height="34dp"
-    android:tint="#FAF8F8"
-    android:viewportHeight="24"
-    android:viewportWidth="24"
-    android:width="34dp">
-      
-    <path
-        android:fillColor="@android:color/white"
-        android:pathData="M11,21h-1l1,-7H7.5c-0.58,0 -0.57,-0.32 -0.38,-0.66 0.19,-0.34 0.05,-0.08 0.07,-0.12C8.48,10.94 10.42,7.54 13,3h1l-1,7h3.5c0.49,0 0.56,0.33 0.47,0.51l-0.07,0.15C12.96,17.55 11,21 11,21z"/>
-    
-</vector>
