--- app/src/main/res/drawable/clockblack.xml
@@ -1,12 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:autoMirrored="true"
-    android:height="34dp"
-    android:tint="#101010"
-    android:viewportHeight="24"
-    android:viewportWidth="24"
-    android:width="34dp">
-      
-    <path android:fillColor="@android:color/white"
-        android:pathData="M11.99,2C6.47,2 2,6.48 2,12s4.47,10 9.99,10C17.52,22 22,17.52 22,12S17.52,2 11.99,2zM15.29,16.71L11,12.41V7h2v4.59l3.71,3.71L15.29,16.71z"/>
-    
-</vector>
