--- app/src/main/res/drawable/display.xml
@@ -1,11 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:height="30dp"
-    android:tint="#FAF8F8"
-    android:viewportHeight="24"
-    android:viewportWidth="24"
-    android:width="30dp">
-      
-    <path android:fillColor="@android:color/white"
-        android:pathData="M20,15.31L23.31,12 20,8.69V4h-4.69L12,0.69 8.69,4H4v4.69L0.69,12 4,15.31V20h4.69L12,23.31 15.31,20H20v-4.69zM12,18c-3.31,0 -6,-2.69 -6,-6s2.69,-6 6,-6 6,2.69 6,6 -2.69,6 -6,6z"/>
-    
-</vector>
