--- app/src/main/res/drawable/energy.xml
@@ -1,10 +1,0 @@
-<vector xmlns:android="http://schemas.android.com/apk/res/android"
-    android:height="30dp"
-    android:tint="#FAF8F8"
-    android:viewportHeight="24"
-    android:viewportWidth="24"
-    android:width="30dp">
-      
-    <path android:fillColor="@android:color/white" android:pathData="M12,3c-4.8,0 -9,3.86 -9,9c0,2.12 0.74,4.07 1.97,5.61L3,19.59L4.41,21l1.97,-1.97C7.93,20.26 9.88,21 12,21c2.3,0 4.61,-0.88 6.36,-2.64C20.12,16.61 21,14.3 21,12l0,-9L12,3zM15.83,12.26l-5.16,4.63c-0.16,0.15 -0.41,0.14 -0.56,-0.01c-0.14,-0.14 -0.16,-0.36 -0.04,-0.52l2.44,-3.33l-4.05,-0.4c-0.44,-0.04 -0.63,-0.59 -0.3,-0.89l5.16,-4.63c0.16,-0.15 0.41,-0.14 0.56,0.01c0.14,0.14 0.16,0.36 0.04,0.52l-2.44,3.33l4.05,0.4C15.98,11.41 16.16,11.96 15.83,12.26z"/>
-    
-</vector>
