--- app/src/main/res/drawable/rectangle.xml
@@ -1,15 +1,0 @@
-<!-- rounded_rectangle.xml -->
-<selector xmlns:android="http://schemas.android.com/apk/res/android">
-    <item android:state_pressed="true">
-        <shape android:shape="rectangle">
-            <solid android:color="@color/lightGrey" />
-            <corners android:radius="12dp" />
-        </shape>
-    </item>
-    <item>
-        <shape android:shape="rectangle">
-            <solid android:color="@color/white" />
-            <corners android:radius="12dp" />
-        </shape>
-    </item>
-</selector>
