--- app/src/main/res/layout/fragment_charge.xml
@@ -1,195 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:layout_centerHorizontal="false"
-    android:background="@color/background"
-    android:visibility="visible"
-    tools:context=".Charge"
-    tools:visibility="visible">
-
-    <ImageView
-        android:id="@+id/imageView10"
-        android:layout_width="150dp"
-        android:layout_height="130dp"
-        android:layout_marginLeft="30dp"
-        android:layout_marginTop="300dp"
-        android:src="@drawable/rectangle"
-        android:textColor="@color/lightGrey" />
-
-    <TextView
-        android:id="@+id/textView2"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="30dp"
-        android:layout_marginTop="70dp"
-        android:text="Charge Status"
-        android:textColor="@color/white"
-        android:textSize="24dp" />
-
-    <TextClock
-        android:id="@+id/textClock"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="160dp"
-        android:layout_marginTop="10dp"
-        android:format12Hour="hh:mm a"
-        android:format24Hour="HH:mm"
-        android:textColor="@color/white"
-        android:textSize="20dp" />
-
-    <ImageView
-        android:id="@+id/imageView4"
-        android:layout_width="250dp"
-        android:layout_height="280dp"
-        android:layout_marginLeft="200dp"
-        android:layout_marginTop="20dp"
-        android:src="@drawable/bar2" />
-
-    <TextView
-        android:id="@+id/textView3"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="285dp"
-        android:layout_marginTop="160dp"
-        android:text="88%"
-        android:textColor="@color/white"
-        android:textSize="25dp" />
-
-    <ImageView
-        android:id="@+id/imageView5"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="285dp"
-        android:layout_marginTop="120dp"
-        android:src="@drawable/charge" />
-
-
-    <ImageView
-        android:id="@+id/imageView6"
-        android:layout_width="149dp"
-        android:layout_height="130dp"
-        android:layout_marginLeft="30dp"
-        android:layout_marginTop="150dp"
-        android:src="@drawable/rectangle"
-        android:textColor="@color/lightGrey" />
-
-    <TextView
-        android:id="@+id/textView4"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="38dp"
-        android:layout_marginTop="150dp"
-        android:text="Driving Range"
-        android:textStyle="bold"
-        android:textColor="@color/black"
-        android:textSize="20dp" />
-
-    <TextView
-        android:id="@+id/textView5"
-        android:layout_width="wrap_content"
-        android:layout_height="146dp"
-        android:layout_marginLeft="100dp"
-        android:layout_marginTop="210dp"
-        android:text="436km"
-        android:textSize="20dp"
-        android:textColor="@color/black"
-        android:textStyle="bold" />
-
-    <ImageView
-        android:id="@+id/imageView8"
-        android:layout_width="40dp"
-        android:layout_height="40dp"
-        android:layout_marginLeft="50dp"
-        android:layout_marginTop="210dp"
-        android:src="@drawable/blackcar" />
-
-    <TextView
-        android:id="@+id/textView10"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="38dp"
-        android:layout_marginTop="305dp"
-        android:text="Time for Charge"
-        android:textStyle="bold"
-        android:textColor="@color/black"
-        android:textSize="19dp" />
-
-    <ImageView
-        android:id="@+id/imageView9"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="45dp"
-        android:layout_marginTop="345dp"
-        android:src="@drawable/clockblack" />
-
-    <TextView
-        android:id="@+id/textView11"
-        android:layout_width="wrap_content"
-        android:layout_height="146dp"
-        android:layout_marginLeft="100dp"
-        android:layout_marginTop="350dp"
-        android:text="35mins"
-        android:textSize="20dp"
-        android:textColor="@color/black"
-        android:textStyle="bold" />
-
-    <ImageView
-        android:id="@+id/imageView11"
-        android:layout_width="232dp"
-        android:layout_height="260dp"
-        android:layout_marginTop="400dp"
-        android:src="@drawable/cartra" />
-
-    <TextView
-        android:id="@+id/textView6"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="300dp"
-        android:layout_marginLeft="220dp"
-        android:textColor="@color/white"
-        android:textAlignment="center"
-        android:textSize="22dp"
-        android:text="Charge Level:
-        II" />
-
-    <ImageView
-        android:id="@+id/imageView12"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="300dp"
-        android:layout_marginLeft="200dp"
-        android:src="@drawable/charge" />
-
-    <TextView
-        android:id="@+id/textView7"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="390dp"
-        android:layout_marginLeft="215dp"
-        android:textColor="@color/white"
-        android:textAlignment="center"
-        android:textSize="22dp"
-        android:text="Distance Travelled:
-        236km   " />
-
-    <ImageView
-        android:id="@+id/imageView13"
-        android:layout_width="wrap_content"
-        android:layout_height="wrap_content"
-        android:layout_marginTop="395dp"
-        android:layout_marginLeft="190dp"
-        android:src="@drawable/baseline_fmd_good_24" />
-
-    <Button
-        android:id="@+id/button7"
-        android:layout_width="155dp"
-        android:layout_height="wrap_content"
-        android:layout_marginLeft="240dp"
-        android:layout_marginTop="490dp"
-        android:text="Charging Station"
-        android:textSize="20dp" />
-
-</FrameLayout>
