+++ app/src/main/res/values/colors.xml
@@ -3,14 +3,8 @@
     <color name="black">#FF000000</color>
     <color name="white">#FFFFFFFF</color>
     <color name="background">#404042</color>
-
-    <color name="accent2">#471bfa</color>
     <color name="navbar">#404042</color>
     <color name="icons">#b5b5b3</color>
     <color name="accent">#009DFF</color>
-    <color name="transparent">#00FFFFFF</color>
-    <color name="lightGrey">#d2d6d3</color>
-    <color name="accent">#009DFF</color>
-
 
 </resources>
